requirements.txt
hydra-core
numpy
pandas
scikit-learn
scipy
torch
torchvision
wcmatch